fix: unify completed-event stores and harden stage progression
Event outcome markers (a choice's nextSceneId, e.g. confession_accepted)
were written only to state.completedEvents, but trigger conditions were
checked against the DB table which only holds bare event ids — so an
event gated on an outcome marker never saw it and mis-triggered.

checkEvent now reads the union of state.completedEvents and the DB record
ids, so gating sees every marker. The pure matching logic is split into
event-matching.ts (no storage dependency) so it can be unit-tested, with
events.ts re-exporting it alongside the storage-backed API.

Stage-gating markers are now applied synchronously before the durable DB
write in both pages, so a failed IndexedDB write can no longer silently
strand a hard-won stage unlock.
